input: synaptics: Ignore keypad presses when touchscreen is in use [2/2] When the user's finger is on the touchscreen, don't process any keypad button presses, as an unintended button press will cause poor UX.
This ignores home-button presses (fpc1020) and menu/back button
presses (s1302) when the touchscreen (s3320) detects one or more fingers.

Although this is racy, no locks are necessary because synchronization is
not critical for this feature. In order to fix a bug that existed in a
previous implementation (keypad wouldn't work after the screen was turned
on), touch_active is reset when the screen is turned off.

What I mostly modified is the int_touch() function as I changed it
fully into unsigned char to let it give a valid value of finger_num,
which will be used to identify if there is touch events. But note that
under 2 error conditions, It would not return -ENIVAL, and just return
0 instead. This should be safe though.

@ShreyanshLodha : Thanks to @aviraxp for pointing out the mistake I made while porting the code.
Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Shreyansh Lodha <slodha96@gmail.com>